+++ BGL/doc/BGL/CGAL/Triangulation_vertex_base_with_id_2.h
@@ -1,7 +1,7 @@
 namespace CGAL {
 
 /*!
-\ingroup PkgBGLHelper
+\ingroup BGLGraphExternalIndices
 
 The class `Triangulation_vertex_base_with_id_2` is a model of the
 concept `TriangulationVertexBase_2`, the base vertex of a
@@ -9,7 +9,8 @@ concept `TriangulationVertexBase_2`, the base vertex of a
 index vertices for \sc{Bgl} algorithms.
 
 Note that the user is in charge of setting indices correctly before
-running a graph algorithm.
+running a graph algorithm, by calling the function
+`CGAL::set_triangulation_ids(Triangulation&)`.
 
 \tparam TriangulationTraits_2 is the geometric traits class
 and must be a model of `TriangulationTraits_2`.