+++ Cartesian_kernel/include/CGAL/Cartesian/Point_3.h
@@ -24,6 +24,7 @@ namespace CGAL {
 template < class R_ >
 class PointC3
 {
+  typedef PointC3<R_>                       Self;
   typedef typename R_::Vector_3             Vector_3;
   typedef typename R_::Point_3              Point_3;
   typedef typename R_::Aff_transformation_3 Aff_transformation_3;
@@ -46,6 +47,15 @@ public:
 
   PointC3(const FT &x, const FT &y, const FT &z, const FT &w)
     : base(x, y, z, w) {}
+
+  friend void swap(Self& a, Self& b)
+#ifdef __cpp_lib_is_swappable
+    noexcept(std::is_nothrow_swappable_v<Vector_3>)
+#endif
+  {
+    using std::swap;
+    swap(a.base, b.base);
+  }
 
   const FT & x() const
   {