+++ Installation/CHANGES.md
@@ -21,6 +21,10 @@ Release date: June 2020
 ### Intersecting Sequences of dD Iso-oriented Boxes 
  -   Added parallel versions of the functions `CGAL::box_intersection_d()` and `CGAL::box_self_intersection_d()`.
 
+### CGAL and the Boost Graph Library (BGL)
+ -   Introduced the function `set_triangulation_ids(Triangulation& tr)` which must be used to initialize vertex,
+     edge, and face indices of a triangulation meant to be used with BGL algorithms.
+
 ### Polygon Mesh Processing
 
 -   Introduced a new function, `CGAL::Polygon_mesh_processing::remove_connected_components_of_negligible_size()`, 
@@ -91,6 +95,11 @@ Release date: June 2020
 ### STL Extensions for CGAL
  -   Added a new concurrency tag: `CGAL::Parallel_if_available_tag`. This tag is a convenience typedef to `CGAL::Parallel_tag`
      if the third party library TBB has been found and linked with, and to `CGAL::Sequential_tag` otherwise.
+     
+
+### Convex_hull_3
+- A new overload for `convex_hull_3()` that takes a model of `VertexListGraph` has been added.
+
 
 [Release 5.0](https://github.com/CGAL/cgal/releases/tag/releases%2FCGAL-5.0)
 -----------